Moved code for chat service into separate file
